test_generator: Handle file I/O failures
std::fopen and std::fwrite can both fail (file unable to be opened, or
contents can't be written entirely).

In the event a file cannot be opened successfully, this would crash the
test generator, which is probably not wanted.

Instead, return whether or not generation was successful, and log to
stderr if any errors occur, then exit the program gracefully.